backend/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
watchdog
guessit
tmdbsimple
sqlalchemy
aiofiles
loguru
rich
requests
python-dotenv
pytest
httpx
pytest-asyncio
ffmpeg-python
